[KAN-70] rename: 로그인 회원가입 로직 파일 이돌
